camera_calibration/requirements.txt
av==16.0.1
caliscope==0.8.1
coloredlogs==15.0.1
flatbuffers==25.12.19
fonttools==4.62.1
humanfriendly==10.0
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.23.2
opencv-python==4.13.0.92
packaging==26.2
pandas==2.3.3
pillow==12.2.0
platformdirs==4.9.6
protobuf==7.34.1
pyarrow==24.0.0
Pygments==2.20.0
PySide6_Addons==6.11.0
PySide6_Essentials==6.11.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
rich==15.0.0
rtoml==0.13.0
scipy==1.15.3
shiboken6==6.11.0
six==1.17.0
sympy==1.14.0
tzdata==2026.2
urllib3==2.6.3
